+++ admin/products.php
@@ -1,61 +1,17 @@
 <?php
 require_once "../DB/dbcon.php";
+
+require_once "../controller/ProductsController.php";
+require_once "../controller/BrandController.php";
+require_once "../controller/CategoryController.php";
+
 require "../includes/layout/backHeader.php";
 ?>
-
-<?php
-$dbCon = dbCon($user, $pass);
-$query = $dbCon->prepare("SELECT * FROM product ORDER BY id DESC");
-$query->execute();
-$getProducts = $query->fetchAll();
-
-$queryCat = $dbCon->prepare("SELECT * FROM category");
-$queryCat->execute();
-$getCat = $queryCat->fetchAll();
-
-$queryBrand = $dbCon->prepare("SELECT * FROM Brand");
-$queryBrand->execute();
-$getBrand = $queryBrand->fetchAll();
-?>
-
-<head>
-
-    <style>
-        .content {
-            width: 90%;
-            display: none;
-        }
-
-        .active {
-            display: block;
-        }
-
-        .nav-tabs {
-            list-style-type: none;
-            padding: 5px;
-        }
-
-        .nav-tabs li {
-            display: inline;
-            padding: 5px 10px;
-            background: lightgrey;
-            cursor: pointer;
-        }
-
-        li.active-tab {
-            background: lightblue;
-        }
-    </style>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
-</head>
 
 <div class="container">
     <ul class="nav-tabs">
         <li class="active-tab" data-link="page-1">All Products</li>
         <li data-link="page-2">New Product</li>
-        <!-- <li data-link="page-3">Third Page</li>
-        <li data-link="page-4">Fourth Page</li> -->
     </ul>
     <div class="content active" id="page-1">
     <div class="column">
@@ -64,7 +20,6 @@ $getBrand = $queryBrand->fetchAll();
             <table class="highlight">
                 <thead>
                     <tr>
-                        <th>productID</th>
                         <th>title</th>
                         <th>description</th>
                         <th>price</th>
@@ -75,24 +30,26 @@ $getBrand = $queryBrand->fetchAll();
                 </thead>
                 <tbody>
                     <?php
-                     foreach ($getProducts as $products) {
-                        echo "<tr>";
-                        echo "<td>" . $products['id'] . "</td>";
-                        echo "<td>" . $products['title'] . "</td>";
-                        echo "<td>" . $products['description'] . "</td>";
-                        echo "<td>" . $products['price'] . "</td>";
-                        echo "<td>" . "<img src='../crud/products/img/" . $products['image'] . "' width='120' height='120' alt='images'>" . "</td>";
-                        echo "<td>";
+                     $products = new ProductsController;
+                     $result = $products->all();
+                     if ($result) {
+                        foreach ($result as $row) {
+                            echo "<tr>";
+                            echo "<td>" . $row['title'] . "</td>";
+                            echo "<td>" . $row['description'] . "</td>";
+                            echo "<td>" . $row['price'] . "</td>";
+                            echo "<td>" . "<img src='../crud/products/img/" . $row['image'] . "' width='120' height='120' alt='images'>" . "</td>";
+                            echo "<td>";
 
-                        echo "</td>";
-                        echo '<td><a href="editProduct.php?id=' . $products['id'] . '" class="waves-effect waves-light btn" ">Edit</a></td>';
-                        echo '<td><a href="../crud/products/deleteProduct.php?id=' . $products['id'] . '" class="waves-effect waves-light btn red" onclick="return confirm(\'Delete! are you sure?\')">Delete</a></td>';
-                        echo "</tr>";
+                            echo "</td>";
+                            echo '<td><a href="editProduct.php?id=' . $row['id'] . '" class="waves-effect waves-light btn" ">Edit</a></td>';
+                            echo '<td><a href="../crud/products/deleteProduct.php?id=' . $row['id'] . '" class="waves-effect waves-light btn red" onclick="return confirm(\'Delete! are you sure?\')">Delete</a></td>';
+                            echo "</tr>";
+                        }
                      }
                     ?>
                 </tbody>
             </table>
-
         </div>
     </div>
 
@@ -126,28 +83,43 @@ $getBrand = $queryBrand->fetchAll();
                     <!-- RELATED CATEGORY -->
                     <div class="input-field">
                         <label class="w-100 p-1" for="">Category</label>
-                        <?php foreach ($getCat as $getCat) {
-                            echo "<input class='' type='checkbox' id='" . $getCat['title'] . "' name='category[]' value='" . $getCat['cat_id'] . "'>";
-                            echo "<label class='p-1' for='" . $getCat['title'] . "'>" . $getCat['title'] . "</label>";
+                        <?php
+                            $category = new CategoryController;
+                            $result = $category->all();
+                            if ($result) {
+                                foreach ($result as $row) {
+                            echo "<input class='' type='checkbox' id='" . $row['title'] . "' name='category[]' value='" . $row['cat_id'] . "'>";
+                            echo "<label class='p-1' for='" . $row['title'] . "'>" . $row['title'] . "</label>";
                         }
-                        ?>
+                    }
+                    ?>
                     </div>
                     <!-- RELATED BRANDS -->
                     <div class="input-field">
                         <label class="w-100 p-1" for="">Brand</label>
-                        <?php foreach ($getBrand as $getBrand) {
-                            echo "<input type='checkbox' id='" . $getBrand['title'] . "' name='brand[]' value='" . $getBrand['brand_id'] . "'>";
-                            echo "<label class='p-1' for='" . $getBrand['title'] . "'>" . $getBrand['title'] . "</label>";
+                        <?php
+                            $brands = new BrandController;
+                            $result = $brands->all();
+                            if ($result) {
+                                foreach ($result as $row) {
+                            echo "<input type='checkbox' id='" . $row['title'] . "' name='brand[]' value='" . $row['brand_id'] . "'>";
+                            echo "<label class='p-1' for='" . $row['title'] . "'>" . $row['title'] . "</label>";
                         }
-                        ?>
+                    }
+                    ?>
                     </div>
                     <!-- RELATED PRODUCT -->
                     <div class="input-field">
                         <label class="w-100 p-1" for="">Related Products</label>
-                        <?php foreach ($getProducts as $product) {
-                            echo "<input type='checkbox' id='" . $product['title'] . "' name='product[]' value='" . $product['id'] . "'>";
-                            echo "<img src='../crud/products/img/" . $product['image'] . "' width='120' height='120' alt='images'>";
-                            echo "<label class='p-1' for='" . $product['title'] . "'>" . $product['title'] . "</label>";
+                        <?php
+                            $products = new ProductsController;
+                            $result = $products->all();
+                            if ($result) {
+                                foreach ($result as $row) {
+                            echo "<input type='checkbox' id='" . $row['title'] . "' name='product[]' value='" . $row['id'] . "'>";
+                            echo "<img src='../crud/products/img/" . $row['image'] . "' width='120' height='120' alt='images'>";
+                            echo "<label class='p-1' for='" . $row['title'] . "'>" . $row['title'] . "</label>";
+                            }
                         }
                         ?>
                     </div>
@@ -157,29 +129,6 @@ $getBrand = $queryBrand->fetchAll();
             </form>
         </div>
     </div>
-
-    <!-- <div class="content" id="page-3">
-        This is going to be the content for the fourth page. This block isn't even visible until jQuery and user
-        action makes it so.
-    </div>
-
-    <div class="content" id="page-4">
-        This is going to be the content for the fourth page. This block isn't even visible until jQuery and user
-        action makes it so. </div> -->
-
 </div>
-<script>
-    // page is ready to go via HTML, jquery only comes into play when user interacts. 
-    $(".nav-tabs li").click(function () {
-        // only executes if tab/page not active
-        if (!$(this).hasClass('active-tab')) {
-            $(".nav-tabs li").removeClass("active-tab");
-            $(this).addClass('active-tab');
-            var page = "#" + $(this).attr("data-link");
-            $(".content").removeClass('active');
-            $(page).addClass('active');
-        }
-    });
-</script>
 
 <?php require "../includes/layout/backFooter.php"; ?>
